+++ fixtures/contracts/Basic.sol
@@ -28,23 +28,11 @@ contract Basic {
 		return boolValue;
 	}
 
-	function getValues()
-		public
-		view
-		returns (
-			uint256,
-			string memory,
-			bool
-		)
-	{
+	function getValues() public view returns (uint256, string memory, bool) {
 		return (intValue, stringValue, boolValue);
 	}
 
-	function setValues(
-		uint256 _value,
-		string memory _stringValue,
-		bool _boolValue
-	) public {
+	function setValues(uint256 _value, string memory _stringValue, bool _boolValue) public {
 		intValue = _value;
 		stringValue = _stringValue;
 		boolValue = _boolValue;
@@ -62,19 +50,11 @@ contract Basic {
 		revert('REVERTED WITH REVERT');
 	}
 
-	function firesMultiValueEvent(
-		string memory str,
-		uint256 val,
-		bool flag
-	) public {
+	function firesMultiValueEvent(string memory str, uint256 val, bool flag) public {
 		emit MultiValueEvent(str, val, flag);
 	}
 
-	function firesMultiValueIndexedEvent(
-		string memory str,
-		uint256 val,
-		bool flag
-	) public {
+	function firesMultiValueIndexedEvent(string memory str, uint256 val, bool flag) public {
 		emit MultiValueIndexedEvent(str, val, flag);
 	}