+++ packages/web3-eth-contract/src/utils.ts
@@ -16,8 +16,13 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { Web3ContractError } from 'web3-errors';
-import { AbiFunctionFragment } from 'web3-eth-abi';
-import { TransactionWithSenderAPI, TransactionCall, HexString } from 'web3-types';
+import {
+	TransactionForAccessList,
+	AbiFunctionFragment,
+	TransactionWithSenderAPI,
+	TransactionCall,
+	HexString,
+} from 'web3-types';
 import { isNullish, mergeDeep, toHex } from 'web3-utils';
 import { encodeMethodABI } from './encoding';
 import {
@@ -25,6 +30,7 @@ import {
 	PayableCallOptions,
 	ContractOptions,
 	Web3ContractContext,
+	ContractInitOptions,
 } from './types';
 
 export const getSendTxParams = ({
@@ -133,12 +139,59 @@ export const getEstimateGasParams = ({
 	return txParams as TransactionWithSenderAPI;
 };
 
-export const isContractInitOptions = (options: unknown): options is ContractOptions =>
+export const isContractInitOptions = (options: unknown): options is ContractInitOptions =>
 	typeof options === 'object' &&
 	!isNullish(options) &&
-	['data', 'from', 'gas', 'gasPrice', 'gasLimit', 'address', 'jsonInterface'].some(
-		key => key in options,
-	);
+	[
+		'data',
+		'from',
+		'gas',
+		'gasPrice',
+		'gasLimit',
+		'address',
+		'jsonInterface',
+		'syncWithContext',
+	].some(key => key in options);
 
 export const isWeb3ContractContext = (options: unknown): options is Web3ContractContext =>
 	typeof options === 'object' && !isNullish(options) && !isContractInitOptions(options);
+
+export const getCreateAccessListParams = ({
+	abi,
+	params,
+	options,
+	contractOptions,
+}: {
+	abi: AbiFunctionFragment;
+	params: unknown[];
+	options?: PayableCallOptions | NonPayableCallOptions;
+	contractOptions: ContractOptions;
+}): TransactionForAccessList => {
+	if (!options?.to && !contractOptions.address) {
+		throw new Web3ContractError('Contract address not specified');
+	}
+
+	if (!options?.from && !contractOptions.from) {
+		throw new Web3ContractError('Contract "from" address not specified');
+	}
+
+	let txParams = mergeDeep(
+		{
+			to: contractOptions.address,
+			gas: contractOptions.gas,
+			gasPrice: contractOptions.gasPrice,
+			from: contractOptions.from,
+			data: contractOptions.data,
+		},
+		options as unknown as Record<string, unknown>,
+	) as unknown as TransactionForAccessList;
+
+	if (!txParams.data || abi.type === 'constructor') {
+		txParams = {
+			...txParams,
+			data: encodeMethodABI(abi, params, txParams.data as HexString),
+		};
+	}
+
+	return txParams;
+};