+++ packages/web3/src/providers.exports.ts
@@ -15,8 +15,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-export type ReconnectOptions = {
-	autoReconnect: boolean;
-	delay: number;
-	maxAttempts: number;
-};
+export { Eip1193Provider, SocketProvider } from 'web3-utils';
+
+export * as http from 'web3-providers-http';
+export * as ipc from 'web3-providers-ipc';
+export * as ws from 'web3-providers-ws';