+++ packages/web3-eth/CHANGELOG.md
@@ -90,3 +90,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Enable transaction with local wallet index in the `to` field (#5731)
 
 ## [Unreleased]
+
+### Changed
+
+-   Update imports statements for objects that was moved between web3 packages (#5771)
+
+### Added
+
+-   Added `createAccessList` functionality ( #5780 )